+++ tests/c_api/test_c_api.nim
@@ -1,29 +1,9 @@
-import lazy_rest_c_api
+import lazy_rest_c_api, c_test, lazy_rest
 
-## Test for the C API of lazy_rest.
-##
-## Don't be fooled, despite this being a Nimrod program, it actually contains
-## little Nimrod code. The single proc of this test just emits lots of C code.
-## The reason for testing C code through the emit pragma is that it leaves us a
-## cross platform portable way of running and invoking C compilers leveraging
-## the work that has already been done for Nimrod itself.
-##
-## Since writing all C code embedded inside the emit pragma is quite painful we
-## use slurp the get real C code from an external file which allows our puny
-## editors to syntax highlight it properly.
-
-const
-  c_source = slurp("test_c_api.c")
-  error_template = slurp("../errors/custom_default_error.rst")
-  special_options = slurp("../runtime_change/special.cfg")
-
-{.emit:c_source.}
-
-proc run_tests() =
-  let
-    e = error_template.cstring
-    o = special_options.cstring
-  {.emit:"""run_c_test(`e`, `o`);""".}
+proc ignore_msg_handler*(filename: string, line, col: int,
+    msgkind: TMsgKind, arg: string): string {.procvar, raises: [], exportc.} =
+  ## Custom handler to ignore all errors, used in the C source code.
+  discard
 
 when isMainModule:
   run_tests()