+++ tests/include/test_include.nim
@@ -1,17 +1,34 @@
-import lazy_rest
+import lazy_rest, os
 
-when isMainModule:
-  echo "Testing the first one, this should be easy."
-  writeFile("simple_include_01.html",
-    safe_rst_file_to_html("simple_include_01.rst"))
+const
+  rst_files = [
+    "simple_include_01", "recursion_base", "complex_recursion_a",
+    "relative", "absolute_unix", "absolute_windows", "repeat",
+    ]
+  flattened_files = ["flattened_includes"]
 
-  echo "Ok, now testing simple recursion, use Ctrl+C to escape…"
-  writeFile("recursion_base.html",
-    safe_rst_file_to_html("recursion_base.rst"))
 
-  echo "Ok, now the complex trick, cyclick recursion…"
-  writeFile("complex_recursion_a.html",
-    safe_rst_file_to_html("complex_recursion_a.rst"))
+proc flattened_find_file(ignored, target_filename: string):
+    string {.procvar, raises:[].} =
+  let target = target_filename.extract_filename
+  if target.exists_file:
+    result = target
 
-  echo "Testing repetitive relative includes…"
-  writeFile("relative.html", safe_rst_file_to_html("relative.rst"))
+
+proc test() =
+  # Normal includes.
+  for name in rst_files:
+    let
+      src = name & ".rst"
+      dest = name & ".html"
+    dest.write_file(src.safe_rst_file_to_html)
+
+  # Flattened ones.
+  for name in flattened_files:
+    let
+      src = name & ".rst"
+      dest = name & ".html"
+    dest.write_file(src.safe_rst_file_to_html(find_file = flattened_find_file))
+
+
+when isMainModule: test()